refactor(state): move message in channel state
